Even stricter verification of local copy
We must be absolutely sure that the local copy is consistent.

A check between packed and observed version before the attempted copy
raises confidence that the array pointer is still valid and gives us a
way of aborting the current attempt before having to perform the entire
copy.

After the copy is completed, perform a three-way check to ensure the
global, observed, and local versions are all equal.

The symptom that led to these changes was that local array copies were
obviously being made of incompletely updated source arrays since null
pointers were being pushed into the blocks vector.