Ensure random distribution covers entire range
This seemingly innocent bug has huge impact on results. The constructor
defaults to a minimum of 0, which meant that the sign bit was
deterministically left unset. This led to a larger ratio of deletes
being performed (1/64 ~ 1.6 % more likely), which in turn led to a
smaller queue, better performance, and additional spy() invocations.