Add implementation of the contention avoiding priority queue
The Contention Avoiding Priority Queue (CA-PQ) is a relaxed priority
queue described in the paper "The Contention Avoiding Concurrent
Priority Queue" which is published in the 29th International Workshop
on Languages and Compilers for Parallel Computing (LCPC 2016). A
preprint version of the paper can be found at
https://www.it.uu.se/research/group/languages/software/ca_pq . This
commit adds CA-PQ as well as the priority queues CA-DM, CA-IN and
CATree to the random benchmark. CA-DM, CA-IN and CATree are derived
from CA-PQ by turning off features and are all described in the paper
mentioned above.