Update the SprayList and the Linden priority queue
The code for the SprayList and the Linden priority queue is updated to
correspond to a patched version of the latest code from the SprayList
repository (*). In contrast to the previous version of the Linden
priority queue implementation, the version included in this commit
supports duplicated keys. In contrast to the previous version of the
SprayList implementation, the version included in this commit supports
associating a value with the keys.

The patch that has been applied on top of the latest SprayList version
(*) improves SprayList's memory management. (A pull request will soon
be submitted to the SprayList repository with this patch.)

Without the patch it is only possible to specify how much memory each
thread will be able to allocate. This is problematic for the single
source shortest paths benchmark as it makes it impossible to run the
benchmark with a single thread and many threads using the same
binary. The benchmark would either run out of memory when a single
thread is used or there would not be enough memory on the machine when
many threads are used.

With the patch one can specify how much memory all threads can
allocated together (the #define SSALLOC_SIZE_ALL in
include/ssalloc.h). The patch makes sure that this amount of memory is
divided between the threads (Note that the global variable
number_of_threads has to be set to the correct value for this to
work).

(*) https://github.com/jkopinsky/SprayList/tree/e6c851d93156646efa4ce0a9dbdb2e1a3e8d3693