Add delete_min method that returns both key and value
This commit adds a delete_min method that returns both the key and the
value to the priority queues that don't already have such a method. It
is useful to get both the key and the value from a deleted item in
algorithms such as Dijkstra's single source shortest paths algorithm.