Add shortest paths benchmark that load graph from file
This shortest path benchmark (file_shortest_paths.cpp) is implemented
as a new benchmark instead of as a new option in the already existing
shortest paths benchmark (shortest_paths.cpp) that generates a random
graph instead of loading a graph from a file. The reason for this is
that the new benchmark is a heavily modified version of the old
version. Task allocation and the global counter num_tasks is avoided
in the new benchmark. Furthermore the verify_graph function is not
present as the calculated paths can be verified by verifying the
output file. The implementation of this benchmark is inspired by the
sssp benchmark in the SprayList repository (See
lib/spraylist_linden/sssp.c).

Additionally this commit contains a program that can be used to
generate input graphs for the benchmark (generate_random_graph.cpp).